added modelkey field to regression.py and moved jade.py from the bin to the transform directory.
